+++ src/config/plugin.ts
@@ -112,7 +112,7 @@ export class Plugin implements IPlugin {
 
   manifest!: Manifest
 
-  commands!: Command.Plugin[]
+  commands!: Command.Loadable[]
 
   hooks!: {[k: string]: string[]}
 
@@ -157,9 +157,10 @@ export class Plugin implements IPlugin {
     this.hooks = mapValues(this.pjson.oclif.hooks || {}, i => Array.isArray(i) ? i : [i])
 
     this.manifest = await this._manifest(Boolean(this.options.ignoreManifest), Boolean(this.options.errorOnManifestCreate))
-    this.commands = Object.entries(this.manifest.commands)
+    this.commands = Object
+    .entries(this.manifest.commands)
     .map(([id, c]) => ({...c, pluginAlias: this.alias, pluginType: this.type, load: async () => this.findCommand(id, {must: true})}))
-    this.commands.sort((a, b) => a.id.localeCompare(b.id))
+    .sort((a, b) => a.id.localeCompare(b.id))
   }
 
   get topics(): Topic[] {
@@ -191,7 +192,8 @@ export class Plugin implements IPlugin {
       const p = path.parse(file)
       const topics = p.dir.split('/')
       const command = p.name !== 'index' && p.name
-      return [...topics, command].filter(f => f).join(':')
+      const id = [...topics, command].filter(f => f).join(':')
+      return id === '' ? '.' : id
     })
     this._debug('found commands', ids)
     return ids