+++ src/help/util.ts
@@ -2,6 +2,7 @@ import * as ejs from 'ejs'
 import {Config as IConfig, HelpOptions} from '../interfaces'
 import {Help, HelpBase} from '.'
 import ModuleLoader from '../module-loader'
+import {collectUsableIds} from '../config/util'
 
 interface HelpBaseDerived {
   new(config: IConfig, opts?: Partial<HelpOptions>): HelpBase;
@@ -38,32 +39,27 @@ export function template(context: any): (t: string) => string {
 function collateSpacedCmdIDFromArgs(argv: string[], config: IConfig): string[] {
   if (argv.length === 1) return argv
 
-  const ids = new Set(config.commandIDs.concat(config.topics.map(t => t.name)))
-
   const findId = (argv: string[]): string | undefined => {
+    const ids = collectUsableIds(config.commandIDs)
+
     const final: string[] = []
-    const idPresent = (id: string) => ids.has(id)
+    const idPresent = (id: string) => ids.includes(id)
     const isFlag = (s: string) => s.startsWith('-')
     const isArgWithValue = (s: string) => s.includes('=')
     const finalizeId = (s?: string) => s ? [...final, s].join(':') : final.join(':')
 
-    const hasSubCommandsWithArgs = () => {
+    const hasArgs = () => {
       const id = finalizeId()
-      /**
-       * Get a list of sub commands for the current command id. A command is returned as a subcommand under either
-       * of these conditions:
-       * 1. the `id` start with the current command id.
-       * 2. any of the aliases start with the current command id.
-       */
-      const subCommands = config.commands.filter(c => (c.id).startsWith(id) || c.aliases.some(a => a.startsWith(id)))
-      return Boolean(subCommands.find(cmd => cmd.strict === false || cmd.args?.length > 0))
+      if (!id) return false
+      const cmd = config.findCommand(id)
+      return Boolean(cmd && (cmd.strict === false || cmd.args?.length > 0))
     }
 
     for (const arg of argv) {
       if (idPresent(finalizeId(arg))) final.push(arg)
       // If the parent topic has a command that expects positional arguments, then we cannot
       // assume that any subsequent string could be part of the command name
-      else if (isArgWithValue(arg) || isFlag(arg) || hasSubCommandsWithArgs()) break
+      else if (isArgWithValue(arg) || isFlag(arg) || hasArgs()) break
       else final.push(arg)
     }
 
@@ -94,4 +90,10 @@ export function standardizeIDFromArgv(argv: string[], config: IConfig): string[]
   if (config.topicSeparator === ' ') argv = collateSpacedCmdIDFromArgs(argv, config)
   else if (config.topicSeparator !== ':') argv[0] = toStandardizedId(argv[0], config)
   return argv
+}
+
+export function getHelpFlagAdditions(config: IConfig): string[] {
+  const helpFlags = ['--help']
+  const additionalHelpFlags = config.pjson.oclif.additionalHelpFlags ?? []
+  return [...new Set([...helpFlags, ...additionalHelpFlags]).values()]
 }