+++ src/module-loader.ts
@@ -13,7 +13,7 @@ const getPackageType = require('get-package-type')
  * Defines file extension resolution when source files do not have an extension.
  */
 // eslint-disable-next-line camelcase
-const s_EXTENSIONS: string[] = ['.js', '.mjs', '.cjs']
+const s_EXTENSIONS: string[] = ['.ts', '.js', '.mjs', '.cjs']
 
 /**
  * Provides a mechanism to use dynamic import / import() with tsconfig -> module: commonJS as otherwise import() gets
@@ -110,6 +110,9 @@ export default class ModuleLoader {
     case '.js':
       return getPackageType.sync(filePath) === 'module'
 
+    case '.ts':
+      return getPackageType.sync(filePath) === 'module'
+
     case '.mjs':
       return true
 
@@ -139,16 +142,28 @@ export default class ModuleLoader {
     } catch {
       filePath = Config.tsPath(config.root, modulePath)
 
-      // Try all supported extensions.
-      if (!fs.existsSync(filePath)) {
-        // eslint-disable-next-line camelcase
-        for (const extension of s_EXTENSIONS) {
-          const testPath = `${filePath}${extension}`
-
-          if (fs.existsSync(testPath)) {
-            filePath = testPath
-            break
+      let fileExists = false
+      let isDirectory = false
+      if (fs.existsSync(filePath)) {
+        fileExists = true
+        try {
+          if (fs.lstatSync(filePath)?.isDirectory?.()) {
+            fileExists = false
+            isDirectory = true
           }
+        } catch {}
+      }
+
+      if (!fileExists) {
+        // Try all supported extensions.
+        let foundPath = ModuleLoader.findFile(filePath)
+        if (!foundPath && isDirectory) {
+          // Since filePath is a directory, try looking for index file.
+          foundPath = ModuleLoader.findFile(path.join(filePath, 'index'))
+        }
+
+        if (foundPath) {
+          filePath = foundPath
         }
       }
 
@@ -156,5 +171,25 @@ export default class ModuleLoader {
     }
 
     return {isESM, filePath}
+  }
+
+  /**
+   * Try adding the different extensions from `s_EXTENSIONS` to find the file.
+   *
+   * @param {string} filePath - File path to load.
+   *
+   * @returns {string | null} Modified file path including extension or null if file is not found.
+   */
+  static findFile(filePath: string) : string | null {
+    // eslint-disable-next-line camelcase
+    for (const extension of s_EXTENSIONS) {
+      const testPath = `${filePath}${extension}`
+
+      if (fs.existsSync(testPath)) {
+        return testPath
+      }
+    }
+
+    return null
   }
 }