+++ test/cli-ux/styled/table.test.ts
@@ -319,3 +319,4 @@ describe('styled/table', () => {
     })
   })
 })
+